(with Philippe) changes to allow bare strings (and other things) as products. Expand test

+++ test/Integration/IntTestAnalyzer_module.cc
@@ -3,7 +3,7 @@
 #include "test/TestObjects/ToyProducts.h"
 
 namespace arttest {
-   typedef GenericOneSimpleProductAnalyzer<double, IntProduct> IntTestAnalyzer;
+   typedef GenericOneSimpleProductAnalyzer<int, IntProduct> IntTestAnalyzer;
 }
 
 DEFINE_ART_MODULE(arttest::IntTestAnalyzer);
